:hammer: create new files for hf data loading and collating
