+++ launch_scripts/eval_captioner.py
@@ -7,7 +7,8 @@ from omegaconf import OmegaConf
 from olmo.eval.inf_evaluator import EvaluatorConfig
 from olmo.train.trainer_config import FSDPConfig, FSDPPrecision
 from olmo.models.model import FSDPWrapStrategy
-from olmo.data.data_loader import DataLoaderConfig, HFDataLoaderConfig
+from olmo.data.data_loader import DataLoaderConfig
+from olmo.hf_train.data_loader import HFDataLoaderConfig
 from olmo.torch_util import get_world_size
 from olmo.util import clean_opt, prepare_torchrun_environment, select_checkpoint
 from scripts.mm_eval import ModelEvaluator, EvalConfig, DatasetEvaluatorConfig
