+++ olmo/hf_train/data_loader.py
@@ -8,67 +8,30 @@ import numpy as np
 import omegaconf
 from torch.utils.data import DataLoader, DistributedSampler
 
-from olmo.config import BaseConfig
-from olmo.data.dataset import DeterministicDataset
 from olmo.data.get_dataset import get_dataset_by_name
 from olmo.data.iterable_dataset_mixture import IterableDatasetMixture
-from olmo.models.molmo.molmo import MolmoConfig
+from olmo.hf_train.collator import HF_MMCollator
+from olmo.data.data_loader import DataLoaderConfig
 from olmo.torch_util import get_global_rank, get_world_size
+from transformers import AutoProcessor
+from olmo.hf_train.dataset import HFDeterministicDataset
 
 log = logging.getLogger(__name__)
 
 
 @dataclass
-class RootSizeMixture(BaseConfig):
-    rate: float
-    mixture: Dict[str, Optional[float]]
-
-
-@dataclass
-class DataLoaderConfig(BaseConfig):
-    """Configuration for a torch `DataLoader`"""
-
-    dataset: Optional[str] = None
-    """Dataset name, will be used for `get_dataset_by_name`"""
-
-    mixture: Optional[Dict[str, float]] = None
-    """Mixture of dataset names and sampling rates"""
-
-    root_size_mixture: Optional[List[RootSizeMixture]] = None
-    """Mixture-of-mixtures where sub-mixtures rates are determined by the root dataset size"""
-
-    split: str = omegaconf.MISSING
-    """Dataset split to load"""
-
-    seed: int = omegaconf.MISSING
-    """Dataset seed for shuffling and augmentation"""
-
-    pad: Optional[str] = "to_max"
-    """How to pad in the collator"""
-
-    sequence_length: Optional[int] = None
-    """Max sequence length to truncate examples to in the Collator"""
-
-    shuffle: Optional[bool] = True
-    """Should the data be shuffled"""
-
-    # DataLoader args
-    num_workers: int = 0
-    drop_last: bool = False
-    pin_memory: bool = True
-    prefetch_factor: Optional[int] = None
-    persistent_workers: bool = False
-    timeout: int = 0
-
+class HFDataLoaderConfig(DataLoaderConfig):
+    """Configuration for a torch `DataLoader` for evaluating an HF model"""
+    
     def build_eval_dataloader(
         self,
-        model_config: MolmoConfig,
         batch_size: int,
         for_inference: bool,
+        preprocessor: AutoProcessor = None,
         include_metadata: bool = None,
         pad_batches: bool = False,
         max_steps_for_padding=None,
-        include_image: bool = False,
+        return_dataset=False
     ) -> DataLoader:
         assert self.mixture is None and self.root_size_mixture is None
         log.info(f"Loading eval dataset: {self.dataset}/{self.split}")
@@ -89,15 +52,15 @@ class DataLoaderConfig(BaseConfig):
                 # exactly the same number of batches
                 n_pad = (n_steps*global_batch_size) - len(dataset)
 
-        max_seq_len = self.sequence_length if self.pad else None
-        preprocessor = model_config.build_preprocessor(
-            for_inference=for_inference, is_training=False, include_image=include_image, max_seq_len=max_seq_len)
-        dataset = DeterministicDataset(
+        dataset = HFDeterministicDataset(
             dataset=dataset,
             seed=self.seed,
             preprocessor=preprocessor,
-            n_pad=n_pad
+            n_pad=n_pad,
+            for_inference=for_inference,
         )
+        if return_dataset:
+            return dataset
         sampler = DistributedSampler(
             dataset,
             drop_last=self.drop_last,
@@ -106,12 +69,11 @@ class DataLoaderConfig(BaseConfig):
             rank=get_global_rank(),
             seed=self.seed,
         )
-
+        
         return DataLoader(
             dataset,
             batch_size=batch_size,
-            collate_fn=model_config.build_collator(
-                self.sequence_length, self.pad, include_metadata=include_metadata),
+            collate_fn=HF_MMCollator(tokenizer=preprocessor.tokenizer),
             num_workers=self.num_workers,
             sampler=sampler,
             pin_memory=self.pin_memory,
@@ -120,9 +82,9 @@ class DataLoaderConfig(BaseConfig):
             timeout=self.timeout,
         )
 
-    def build_train_dataloader(
+    def build_train_dataset(
         self, 
-        model_config: MolmoConfig,
+        preprocessor: AutoProcessor,
         global_batch_size: int, 
         device=None
     ) -> DataLoader:
@@ -165,10 +127,7 @@ class DataLoaderConfig(BaseConfig):
             for ix in np.argsort(rates)[::-1]:
                 log.info(f"{names[ix]}: {100*rates[ix]:0.2f}")
 
-        max_seq_len = self.sequence_length if self.pad else None
-        preprocessor = model_config.build_preprocessor(
-            for_inference=False, is_training=True, max_seq_len=max_seq_len)
-        datasets = [DeterministicDataset(ds, preprocessor, self.seed) for ds in datasets]
+        datasets = [HFDeterministicDataset(ds, preprocessor, self.seed) for ds in datasets]
 
         dataset = IterableDatasetMixture(
             datasets=datasets,
@@ -177,15 +136,5 @@ class DataLoaderConfig(BaseConfig):
             seed=self.seed,
             shuffle=self.shuffle,
         )
-        return DataLoader(
-            dataset,
-            batch_size=dataset.device_batch_size,
-            drop_last=self.drop_last,
-            collate_fn=model_config.build_collator(
-                self.sequence_length, self.pad, include_metadata=False),
-            num_workers=self.num_workers,
-            pin_memory=self.pin_memory,
-            prefetch_factor=None if self.num_workers == 0 else self.prefetch_factor,
-            persistent_workers=self.persistent_workers,
-            timeout=self.timeout,
-        )
+    
+        return dataset
