+++ olmo/hf_train/finetune.py
@@ -30,7 +30,8 @@ from olmo.hf_models import *
 from olmo.hf_models.processing import *
 from launch_scripts.utils import get_hf_evaluation
 
-from olmo.data.data_loader import HFDataLoaderConfig, RootSizeMixture
+from olmo.data.data_loader import RootSizeMixture
+from olmo.hf_train.data_loader import HFDataLoaderConfig
 from olmo.torch_util import get_world_size
 from olmo.eval.loss_evaluator import LossDatasetEvaluatorConfig
 from launch_scripts.utils import VISION_BACKBONES
